WIP Separate config files
